requirements.txt
arxiv==2.1.3
langchain_core==0.2.36
langchain_groq==0.1.9
langchain_huggingface==0.0.3
pandas==2.2.2
python-dotenv==1.0.1
pytz==2024.1
st_tailwind==1.0.1
streamlit==1.38.0
supabase==2.7.3
